+++ src/main/java/com/example/myShop/domain/dto/goods/GoodDto.java
@@ -2,14 +2,13 @@ package com.example.myShop.domain.dto.goods;
 
 import com.example.myShop.domain.dto.category.CategoryDto;
 import com.example.myShop.domain.dto.producer.ProducerDto;
-import com.example.myShop.domain.entity.Category;
 import com.example.myShop.domain.entity.Order;
-import com.example.myShop.domain.entity.Producer;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Value;
 import lombok.extern.jackson.Jacksonized;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static lombok.AccessLevel.PRIVATE;
@@ -26,8 +25,8 @@ import static lombok.AccessLevel.PRIVATE;
 public class GoodDto {
     int id;
     String name;
-    int price;
-    int availability;
+    BigDecimal price;
+    long availability;
     String imageUrl;
 
     ProducerDto producer;
