+++ src/main/java/com/example/myShop/domain/dto/order/OrderDto.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Value;
 import lombok.extern.jackson.Jacksonized;
 
+import java.math.BigDecimal;
+
 import static lombok.AccessLevel.PRIVATE;
 
 /**
@@ -20,7 +22,7 @@ import static lombok.AccessLevel.PRIVATE;
 public class OrderDto {
     int id;
     int count;
-    int price;
+    BigDecimal price;
 
     Goods goods;
     User user;
